+++ crates/router/src/services/api/client.rs
@@ -1,8 +1,9 @@
 use std::time::Duration;
 
+use base64::Engine;
 use error_stack::ResultExt;
 use http::{HeaderValue, Method};
-use masking::PeekInterface;
+use masking::{ExposeInterface, PeekInterface};
 use once_cell::sync::OnceCell;
 use reqwest::multipart::Form;
 use router_env::tracing_actix_web::RequestId;
@@ -10,11 +11,8 @@ use router_env::tracing_actix_web::RequestId;
 use super::{request::Maskable, Request};
 use crate::{
     configs::settings::{Locker, Proxy},
-    consts::LOCKER_HEALTH_CALL_PATH,
-    core::{
-        errors::{ApiClientError, CustomResult},
-        payments,
-    },
+    consts::{BASE64_ENGINE, LOCKER_HEALTH_CALL_PATH},
+    core::errors::{ApiClientError, CustomResult},
     routes::{app::settings::KeyManagerConfig, SessionState},
 };
 
@@ -90,11 +88,11 @@ pub fn create_client(
         (Some(encoded_certificate), Some(encoded_certificate_key)) => {
             let client_builder = get_client_builder(proxy_config, should_bypass_proxy)?;
 
-            let identity = payments::helpers::create_identity_from_certificate_and_key(
+            let identity = create_identity_from_certificate_and_key(
                 encoded_certificate.clone(),
                 encoded_certificate_key,
             )?;
-            let certificate_list = payments::helpers::create_certificate(encoded_certificate)?;
+            let certificate_list = create_certificate(encoded_certificate)?;
             let client_builder = certificate_list
                 .into_iter()
                 .fold(client_builder, |client_builder, certificate| {
@@ -109,6 +107,42 @@ pub fn create_client(
         }
         _ => get_base_client(proxy_config, should_bypass_proxy),
     }
+}
+
+pub fn create_identity_from_certificate_and_key(
+    encoded_certificate: masking::Secret<String>,
+    encoded_certificate_key: masking::Secret<String>,
+) -> Result<reqwest::Identity, error_stack::Report<ApiClientError>> {
+    let decoded_certificate = BASE64_ENGINE
+        .decode(encoded_certificate.expose())
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+
+    let decoded_certificate_key = BASE64_ENGINE
+        .decode(encoded_certificate_key.expose())
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+
+    let certificate = String::from_utf8(decoded_certificate)
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+
+    let certificate_key = String::from_utf8(decoded_certificate_key)
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+
+    let key_chain = format!("{}{}", certificate_key, certificate);
+    reqwest::Identity::from_pem(key_chain.as_bytes())
+        .change_context(ApiClientError::CertificateDecodeFailed)
+}
+
+pub fn create_certificate(
+    encoded_certificate: masking::Secret<String>,
+) -> Result<Vec<reqwest::Certificate>, error_stack::Report<ApiClientError>> {
+    let decoded_certificate = BASE64_ENGINE
+        .decode(encoded_certificate.expose())
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+
+    let certificate = String::from_utf8(decoded_certificate)
+        .change_context(ApiClientError::CertificateDecodeFailed)?;
+    reqwest::Certificate::from_pem_bundle(certificate.as_bytes())
+        .change_context(ApiClientError::CertificateDecodeFailed)
 }
 
 pub fn proxy_bypass_urls(
@@ -246,10 +280,8 @@ impl ProxyClient {
             (Some(certificate), Some(certificate_key)) => {
                 let client_builder =
                     reqwest::Client::builder().redirect(reqwest::redirect::Policy::none());
-                let identity = payments::helpers::create_identity_from_certificate_and_key(
-                    certificate,
-                    certificate_key,
-                )?;
+                let identity =
+                    create_identity_from_certificate_and_key(certificate, certificate_key)?;
                 Ok(client_builder
                     .identity(identity)
                     .build()
